switched to new dependency resolving
by default the resolver resolves in a strict way: it keeps track of all dependants of a package, and if two or more packages have a dependency on the same package it checks if all dependency ranges match, and chooses the latest version within all the matching ranges.

the resolver has a second (currently unused) strategy "latest-compatible": whenever a package requires a version of an already resolved package that isn't api-compatible (according to semver spec) to the resolved one, it up/downgrades to the latest compatible version and checks if that one is still within the required range.

note that the resolver has a lot of methods that are currently unused, will remove them once i'm sure resolving works as expected.